eval/pyproject.toml
[build-system]
requires = [
  "setuptools>=80",
  "setuptools-scm>=8"
]
build-backend = "setuptools.build_meta"

[project]
name = "agentrl-eval"
description = "Evaluation client of AgentRL: Scaling Agentic Reinforcement Learning with a Multi-Turn, Multi-Task Framework"
requires-python = ">=3.10"
license = "MIT"
classifiers = [
  "Operating System :: OS Independent",
  "Programming Language :: Python :: 3",
  "Typing :: Typed"
]
dynamic = ["version"]

dependencies = [
  "anthropic[bedrock]",
  "httpx[http2,socks]",
  "openai~=2.0",
  "polars[async]",
  "Pillow",
  "pydantic-settings"
]

[project.optional-dependencies]
tui = [
  "rich",
  "textual"
]
dev = [
  "pytest",
  "pytest-cov",
  "textual-dev"
]

[project.scripts]
agentrl-eval = "agentrl.eval.__main__:main"

[project.urls]
Repository = "https://github.com/THUDM/AgentRL"
Issues = "https://github.com/THUDM/AgentRL/issues"

[tool.setuptools]
include-package-data = true
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools_scm]
root = ".."
version_scheme = "guess-next-dev"
local_scheme = "node-and-date"
version_file = "src/agentrl/eval/_version.py"
tag_regex = "^eval-v?(?P<version>.+)$"
git_describe_command = ["git", "describe", "--dirty", "--tags", "--long", "--match", "eval-v?*"]
fallback_version = "0.0.0"

[tool.pytest.ini_options]
addopts = "-ra --strict-markers"
pythonpath = [
  "src",
]
testpaths = [
  "tests",
]
